windows/v2-python/requirements.txt
psutil
psycopg[binary]
python-dotenv